Time: 0 ms (100%), Space: 45.2 MB (97.33%) - LeetHub

+++ 3512-minimum-operations-to-make-array-sum-divisible-by-k/3512-minimum-operations-to-make-array-sum-divisible-by-k.cpp
@@ -1,7 +1,7 @@
 class Solution { /// if we noticed we just have to return the remainder after dividing the sum from k (easy - peasy)
 public:
     int minOperations(vector<int>& nums, int k) {
-        long long sum = 0; //intially sum is zero
+        int sum = 0; //intially sum is zero
         for(int i=0; i<nums.size(); i++){
             sum += nums[i]; // update sum till adding last element
         }
